+++ ChangeLog.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## Version [1.0.0](https://github.com/cedx/html.net/compare/v0.10.0...v1.0.0)
-- Added the `New-TextareaElement` and `New-SourceElement` cmdlets.
-- Added the `textarea` and `source` aliases.
+- Added the `New-InputElement` and `New-TextareaElement` cmdlets.
+- Added the `input` and `textarea` aliases.
 
 ## Version [0.10.0](https://github.com/cedx/html.net/compare/v0.9.0...v0.10.0)
 - Added the `New-ButtonElement`, `New-QElement`, `New-SelectElement`, `New-SlotElement` and `New-SourceElement` cmdlets.